requirements.txt
flask
line-bot-sdk
requests
gunicorn
python-dotenv
discord.py
